+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 1.5.2 (...)
 - make sure chosen userId does not exist yet when registering a new account
   (avoiding database constraint exception)
+- script to generate voucher now prints voucherCode
 
 ## 1.5.1 (2018-02-26)
 - remove `addVpnProtoPorts` configuration option, and switch to